# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tenants": {
      "get": {
        "description": "Devuelve una lista de todos los tenants registrados.",
        "operationId": "TenantsController_getTenants",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de tenants recuperada exitosamente."
          }
        },
        "summary": "Obtener todos los tenants",
        "tags": [
          "tenants"
        ]
      },
      "post": {
        "description": "Crea un nuevo tenant con su nombre de negocio y correo electrónico.",
        "operationId": "TenantsController_createTenant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tenant creado exitosamente."
          },
          "400": {
            "description": "Solicitud inválida. Faltan campos obligatorios o son incorrectos."
          }
        },
        "summary": "Crear un nuevo tenant",
        "tags": [
          "tenants"
        ]
      }
    },
...
    "/shifts": {
      "post": {
        "description": "Crea un turno asociado al tenant actual.",
        "operationId": "ShiftsController_createShift",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShiftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Turno creado con éxito."
          },
          "400": {
            "description": "Datos de entrada inválidos."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un nuevo turno",
        "tags": [
          "shifts"
        ]
      }
    },
    "/shifts/{tenantId}/{shiftId}": {
      "put": {
        "description": "Actualiza los datos de un turno existente para el tenant especificado.",
        "operationId": "ShiftsController_updateShift",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shiftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Turno actualizado con éxito."
          },
          "404": {
            "description": "Turno no encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar un turno",
        "tags": [
          "shifts"
        ]
      },
      "delete": {
        "description": "Elimina un turno asociado al tenant especificado.",
        "operationId": "ShiftsController_deleteShift",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shiftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Turno eliminado con éxito."
          },
          "404": {
            "description": "Turno no encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar un turno",
        "tags": [
          "shifts"
        ]
      }
    },
    "/shifts/{shiftId}/close": {
      "patch": {
        "description": "Cierra un turno especificando el efectivo final.",
        "operationId": "ShiftsController_closeShift",
        "parameters": [
          {
            "name": "shiftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseShiftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Turno cerrado con éxito."
          },
          "400": {
            "description": "Datos de entrada inválidos."
          },
          "404": {
            "description": "Turno no encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cerrar un turno",
        "tags": [
          "shifts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Inventory POS API",
    "description": "API para el sistema de inventario y punto de venta",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string",
            "description": "Nombre del negocio del tenant",
            "example": "Tech Corp"
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico del tenant",
            "example": "admin@techcorp.com"
          }
        },
        "required": [
          "businessName",
          "email"
        ]
      },
...
      "CreateShiftDto": {
        "type": "object",
        "properties": {
          "startTime": {
            "type": "string",
            "description": "Hora de inicio del turno en formato ISO 8601",
            "example": "2025-01-21T08:00:00Z"
          },
          "endTime": {
            "type": "string",
            "description": "Hora de finalización esperada del turno en formato ISO 8601",
            "example": "2025-01-21T16:00:00Z"
          },
          "initialCash": {
            "type": "number",
            "description": "Efectivo inicial del turno",
            "example": 100
          }
        },
        "required": [
          "startTime",
          "endTime",
          "initialCash"
        ]
      },
      "CloseShiftDto": {
        "type": "object",
        "properties": {
          "finalCash": {
            "type": "number",
            "description": "Efectivo final al cerrar el turno",
            "example": 120
          }
        },
        "required": [
          "finalCash"
        ]
      }
    }
  }
}